+++ packages/react-traverse-jsx/CHANGELOG.md
@@ -9,4 +9,4 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- `@shopify/react-infuse` package
+- `@shopify/react-traverse-jsx` package